REGRESSION (261912@main): [macOS WK2] imported/w3c/web-platform-tests/css/css-sizing/aspect-ratio/replaced-element-035.html is a flaky failure https://bugs.webkit.org/show_bug.cgi?id=254745 rdar://107423870
Reviewed by Dean Jackson.

Whether we show the elapsed or remaining time label to the right of the scrubber depends on the visibility of the elapsed time label.
Showing the remaining time label will show the "-" symbol prior to the duration while the elapsed time label will not.

To determine whether the elapsed time label should show, we run an "ideal" layout. However, we would account for the existing
state of the elapsed time label's visibility to determine whether to show the elapsed or remaining time label, whereas an ideal
layout should always assume we'd manage to show the elapsed time label.

We re-organize the code to make _performIdealLayout() not account for any elapsed time label visibility state and refactor
the code that performs the layout of the scrubber and time label to the right of the scrubber (duration or remaining) in
a specific _performLayoutWithRightTimeLabel() method which will take in an explicit label parameter.

* LayoutTests/platform/mac-wk2/TestExpectations:
* Source/WebCore/Modules/modern-media-controls/controls/time-control.js:
(TimeControl.prototype.get minimumWidth):
(TimeControl.prototype.get idealMinimumWidth):
(TimeControl.prototype.layout):
(TimeControl.prototype.get _scrubberMargin):
(TimeControl.prototype._performIdealLayout):
(TimeControl.prototype._performLayoutWithRightTimeLabel):

Canonical link: https://commits.webkit.org/262829@main