[Cocoa] font-variation-settings:opsz is not applied when text-rendering:optimizeLegibility is set https://bugs.webkit.org/show_bug.cgi?id=254897 rdar://106814619
Reviewed by Simon Fraser.

text-rendering:optimizeLegibility turns on kCTFontOpticalSizeAttribute:”auto” in Core Text,
and font-variation-settings:opsz turns on kCTFontVariationAttribute:opsz in Core Text. When
both are applied, font-variation-settings is supposed to win. However, in Core Text,
kCTFontOpticalSizeAttribute wins.

The solution for this is to set kCTFontOpticalSizeAttribute to a specific size, given by
font-variation-settings:opsz. This should only happen when OpticalSizingTypes::Everything
is enabled, so this patch adds an optional<float> to that struct. All this is guarded by
a USE() so we can remove it if/when Core text changes to match the CSS behavior.

Test: fast/text/variations/optical-sizing-optimizeLegibility.html

* Source/WTF/wtf/PlatformUse.h:
* Source/WebCore/platform/graphics/FontTaggedSettings.h:
* Source/WebCore/platform/graphics/cocoa/UnrealizedCoreTextFont.cpp:
(WebCore::UnrealizedCoreTextFont::addAttributesForOpticalSizing):
(WebCore::UnrealizedCoreTextFont::modifyFromContext):
* Source/WebCore/platform/graphics/cocoa/UnrealizedCoreTextFont.h:

Canonical link: https://commits.webkit.org/262831@main