Remove Quirk needsMillisecondResolutionForHighResTimeStamp on icourse163 https://bugs.webkit.org/show_bug.cgi?id=254906 rdar://107486680
Reviewed by Brent Fulgham and Simon Fraser.

Previously the site icourse163.org was experiencing a flickering when
clicking on the arrows as explained by the steps to reproduce. The
issue is not happening anymore and the site behaves the same in all
3 browsers. We can remove the quirk.

QA: Zohreh Momeni

* Source/WebCore/dom/ScriptedAnimationController.cpp:
(WebCore::ScriptedAnimationController::serviceRequestAnimationFrameCallbacks):
* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::needsMillisecondResolutionForHighResTimeStamp const): Deleted.
* Source/WebCore/page/Quirks.h:

Canonical link: https://commits.webkit.org/262839@main