Web Inspector: Regression(260175@main) Grid overlay for grids using repeat syntax triggers assert performing downcast https://bugs.webkit.org/show_bug.cgi?id=255311 rdar://107764654
Reviewed by Devin Rousso.

After 260175@main, CSSGridIntegerRepeatValue and CSSGridAutoRepeatValue no longer inherit from CSSValueList, and instead
inherit from CSSValueContainingVector. The inspector overlay code was written in such a way that we were checking if the
class was one of the repeat types, but then downcasting to CSSValueList making that relation less apparent. Converting
the downcasts to dynamicDowncasts where we previously had the `is` checks fixes the issue and guards against this
mistake in the future.

* LayoutTests/inspector/dom/showGridOverlay.html:
* Source/WebCore/inspector/InspectorOverlay.cpp:
(WebCore::authoredGridTrackSizes):

Canonical link: https://commits.webkit.org/262869@main