[iOS] Make eager soft-linking of Data Detection frameworks non-fatal https://bugs.webkit.org/show_bug.cgi?id=255369 <rdar://107752677>
Reviewed by NOBODY (OOPS!).

Eager soft-linking of the Data Detection framework was added in Bug 241267.

Crash telemetry indicates that the soft-linking fails in some cases, perhaps
when WebKit launches while the screen is locked (e.g., for some media uses).

Since this eager-linking was meant to improve performance, and we will perform
the soft-link later if the initial attempt failed, we should just fail gracefully
if the relevant framework directory is not accessible.

This should not introduce any change in behavior, as cases that used to work should
continue. It should simply reduce the changes of a failure terminating a web process
prematurely.

* Source/WebKit/WebProcess/cocoa/WebProcessCocoa.mm: