[UI-side compositing] Hang after dragging scrollbar with scrolling performance testing on https://bugs.webkit.org/show_bug.cgi?id=255365 rdar://107445370
Reviewed by NOBODY (OOPS!).

RemoteScrollingCoordinatorProxy::scrollingTreeNodeDidScroll can be called either after a
scroll in the UI process or after a scrolling tree commit where there was a scroll in the
web process. When we get a scroll from the web process, the tree lock is held in
ScrollingTree::commitTreeState. When we try to access the layout viewport for use in
scrolling performance logging, we try to get the tree lock again, which results in a hang.
To work around this, only grab the tree lock when not commiting the scrolling tree.
Also, make m_inCommitTreeState an atomic bool since it is not protected by m_treeStateLock.
Since std::atomic doesn't have an excplicit copy constructor, we can't use SetForScope.

* Source/WebCore/page/scrolling/ScrollingTree.cpp:
(WebCore::ScrollingTree::layoutViewport const):