Check start after it is modified, not before RegularExpression.match() https://bugs.webkit.org/show_bug.cgi?id=255529
Reviewed by NOBODY (OOPS!).

The first iteration of regex.match is guaranteed to have start be less
than content.length(). This is because of the isEmpty check at the
beginning of the function, which means that the only time start should
be checked is when it changes. This also saves us a function call when
start is at the end of the content.

* Source/JavaScriptCore/inspector/ContentSearchUtilities.cpp:
  (Inspector::ContentSearchUtilities::escapeStringForRegularExpressionSource):
  (Inspector::ContentSearchUtilities::countRegularExpressionMatches):