Correct the state of on-by-default features to 'stable' instead of 'preview' https://bugs.webkit.org/show_bug.cgi?id=255535 <rdar://107159915>
Reviewed by Patrick Angle.

We have a number of features that were in Preview state, that are now on-by-default. Since
Preview is only meant to indicate features that are off-by-default for stable software, we
should reflect the fact that these feature are complete by moving them to 'stable'.

* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:

Canonical link: https://commits.webkit.org/263033@main