(iOS) Dragging the selection in Live Text far away sometimes causes a crash https://bugs.webkit.org/show_bug.cgi?id=255566 rdar://108169157
Reviewed by Wenson Hsieh.

When selecting text with Live Text, if the selection contains different writing modes, a crash
can occur when collecting the selection geometries if the ancestor does not have a renderer.

This PR fixes this by adding a null check for the existence of a renderer.

* Source/WebCore/rendering/RenderObject.cpp:
(WebCore::RenderObject::collectSelectionGeometriesInternal):

Canonical link: https://commits.webkit.org/263053@main